refactor(server): inline five small sibling unit-test modules
Migrate declarative, payload, trust, and pinning sibling *_tests.rs
files into #[cfg(test)] modules in their parents and shrink the
legacy placement inventory (GH-1956).
