+++ crates/harness-server/src/services/execution/mod.rs
@@ -696,8 +696,26 @@ impl ExecutionService for DefaultExecutionService {
 mod continuation_tests;
 
 #[cfg(test)]
-#[path = "../execution_declarative_tests.rs"]
-mod declarative_tests;
+mod declarative_tests {
+    use super::*;
+
+    #[test]
+    fn declarative_submission_requires_a_non_empty_prompt() {
+        for prompt in [None, Some("   ".to_string())] {
+            let request = CreateTaskRequest {
+                definition_id: Some("docs_review".to_string()),
+                prompt,
+                ..Default::default()
+            };
+
+            assert!(matches!(
+                DefaultExecutionService::validate_request(&request),
+                Err(EnqueueTaskError::BadRequest(message))
+                    if message == "declarative workflow submissions require a non-empty prompt"
+            ));
+        }
+    }
+}
 
 #[cfg(test)]
 #[path = "../execution_runtime_policy_tests.rs"]
