+++ crates/harness-server/tests/fixtures/src_test_placement.txt
@@ -58,21 +58,18 @@ file src/http/misc_routes_runtime_tree_tests.rs
 file src/http/startup_tests.rs
 file src/intake/github_issues_rate_limit_tests.rs
 file src/intake/github_issues_tests.rs
-file src/intake/github_issues_trust_tests.rs
 file src/parallel_dispatch_tests.rs
 file src/periodic_reviewer/tests.rs
 file src/project_registry_tests.rs
 file src/quality_trigger_tests.rs
 file src/reconciliation_blocked_done_tests.rs
 file src/reconciliation_local_review_gate_tests.rs
-file src/reconciliation_payload_tests.rs
 file src/reconciliation_ready_to_merge_tests.rs
 file src/reconciliation_state_tests.rs
 file src/reconciliation_tests.rs
 file src/runtime_hosts_tests.rs
 file src/runtime_state_store_tests.rs
 file src/services/execution_continuation_tests.rs
-file src/services/execution_declarative_tests.rs
 file src/services/execution_runtime_policy_tests.rs
 file src/task_db/queries_recovery_tests.rs
 file src/task_queue/aging_tests.rs
@@ -89,13 +86,11 @@ file src/workflow_runtime_submission/declarative_tests.rs
 file src/workflow_runtime_submission/dependency_tests.rs
 file src/workflow_runtime_submission/identity_tests.rs
 file src/workflow_runtime_submission/replay_tests.rs
-file src/workflow_runtime_submission/trust_tests.rs
 file src/workflow_runtime_submission_tests.rs
 file src/workflow_runtime_worker/activity_result_limit_tests.rs
 file src/workflow_runtime_worker/activity_result_tests.rs
 file src/workflow_runtime_worker/prompt_packet/context_provenance_tests.rs
 file src/workflow_runtime_worker/prompt_packet_activity_policy_tests.rs
-file src/workflow_runtime_worker/prompt_packet_pinning_tests.rs
 file src/workflow_runtime_worker/prompt_packet_taint_tests.rs
 file src/workflow_runtime_worker/prompt_packet_tests.rs
 file src/workflow_runtime_worker/turn_engine/stall_tests.rs
